fix(registry): redirect tarball urls to provided registry port+protocol if same host
Ref: https://github.com/npm/npm/pull/9471

npm rewrites tarball URLs to fit the given registry port.
This might seem weird, but it was originally used to allow the registry
to roll out SSL support without needing changes to the registry
contents.

Afterwards, the "feature" was also used by folks running third-party
registries to use http for packuments and https for tarballs, or
vice-versa.

So, we gotta support this. :)