Bring git/revs util up to 100% test coverage.
This uncovered a really interesting bug, and TIL about "peeled tags".

The bug was that, because pacote throws away peeled tags, the logic that
simulates dist-tags.latest in the git ""packument"" was never being hit,
because (a) HEAD was never being returned (as -h and -t were specified,
and HEAD is not a tag or head, but rather a pseudo-head), and (b) the
version _tag_ would never match the SHA, rather the _peeled_ tag would.