Refactor git clone shallow/full logic
This adds the 'gitShallow' option which can be set to false to force a
full fetches, or true to force shallow fetching.

Also rework the clone logic to take a ref object, and parse the revs
itself, ensuring that we always do the most efficient fetch possible,
given the ref provided, and only fall back to a full clone when
absolutely necessary.