Improve integrity consistency and handling
This makes the 'this.integrity' a getter setter that updates the
options, so that the integrity being passed to other parts of the system
is kept in sync.

The setter ensures that we don't ever CHANGE the integrity value (which
could indicate at least a mistake, if not a security exploit), but does
allow for updating with other algorithms, and continually verifying that
the integrity expected for an artifact in one place is also matching
what is expected in another.

Since we have a mix of SHA-1 and SHA-512 in our stack, this always
ensures that we're always using the strongest algorithm available for
integrity verification, but falling back to SHA-1 when SHA-512 is not
available (yet).

Using SHA-512 for make-fetch-happen's caching also increases our cache
hit rate, since we default to that most of the time.