Return the requested spec as the 'from' value
While this is a bit redundant, given the fact that whoever is calling
pacote already knows this by definition, it's also a huge pain to track
when kicking off a ton of requests in parallel.  Much better to just
return it at the end of the process as well, to save the caller that
reassembly and bookkeeping.