git: prefer git+https over git+ssh for hosted repo
This uses the faster and more headless-friendly git+https url when
fetching git remotes on hosted repos, falling back to ssh when https
fails.

Using ssh is better for private repos that require authorization, but
for public repos, https is faster and does not require user interaction
to authorize the use of an SSH key (which may require a passphrase or
2FA).

Resolved url is still reported as the git+ssh url, for consistency.

Fix: #20