registry: verify integrity when loading manifest
I ran into a weird issue where an options object was being reused
improperly between pacote.manifest() calls, leading to the tree builder
believing that it was safe to replace basically any node in the tree
with any new node it fetched.

While this is clearly a bug downstream, it would've been easier to catch
if pacote could flag this as a problem.

As of this change, if the provided integrity does not share any
algorithms in common with the integrity in the dist field (either
dist.integrity or dist.shasum) then it's concatenated onto the provided
integrity.  If it DOES share any algorithms in common, but the hashes
don't match, then an EINTEGRITY error is raised.

This also lets us more easily upgrade to newer integrity algos in the
future, and clients will transparently update their metadata when this
happens.