Extract: rimraf dir contents, not dir itself
That is, without rm -rf'ing all the child nodes underneath it in the
tree.  This supports overwriting a package in the tree with a new
extraction, _without_ deleting all of the package's child nodes in the
node_module heirarchy.

This allows a better tree reification mechanism that does not trigger
EPERM and other folder removal/modification errors, especially on
Windows.

Fix #24