Do not mutate the opts arg passed into Fetcher
This removes a footgun that occurs when the Fetcher objects update their
opts object to include the integrity when it's calculated.  That's an
appropriate thing to do with options objects passed through to subdeps
of pacote, but mutating the object as it was in the *caller* means that
passing the same options object to multiple pacote calls always results
in numerous failures.

CC: @mikemimik