Add packumentCache option
This allows us to avoid hitting the registry repeatedly to get the same
packument multiple times when we are resolving many different specifiers
for the same package name.

Even though these requests are cached after the first hit, just pinging
the registry with an ETag and getting a 304 response still adds up to
quite a bit of time waiting on network requests to complete.

PR-URL: https://github.com/npm/pacote/pull/58
Credit: @isaacs
Close: #58
Reviewed-by: @wraithgar