Preserve git+https auth when provided
This makes pacote use the git+https: url as the resolved value for known
hosts when auth is provided.  When auth is not provided, we store the
ssh url as the resolved value, in order to maintain a canonical save
value that is never git:// for known hosts.

In order to fully fix npm/cli#2054, we will also need to have
@npmcli/arborist store the git+https url in the package-lock.json and
package.json if https auth is provided.

PR-URL: https://github.com/npm/pacote/pull/61
Credit: @isaacs
Close: #61
Reviewed-by: @wraithgar