Use --force when installing deps for git prepare scripts
If a git dependency has conflicting peerDependencies, dependencies with
engines fields that don't match the current environment, and so on, it
isn't much help to crash the installation for these issues.  The user
installing the git dependency is likely not in a position to fix them,
but from npm and Arborist's points of view, it is the 'main' project
that has a problem (since the install is being run as a separate
process).

This commit passes '--force' to the npm child process that installs
dependencies for git packages before running their 'prepare' script, to
ignore all such issues (or reduce them to warnings).

Related to: https://github.com/npm/cli/issues/2548

PR-URL: https://github.com/npm/pacote/pull/63
Credit: @isaacs
Close: #63
Reviewed-by: @nlf