Prevent infinite recursion in git dep preparation
When a git spec is installed, we clone the repo and run 'npm install' if
it has a prepare script, so that the correct bits get loaded.

If one of its dependencies is also a git spec, we do the same thing for
that dependency.

If the dependency graph loops back on itself, we're in trouble.

The solution here is to track the path of git prepare steps we've
followed, and bail out of it when a cycle is detected.  This MAY result
in a package dependency not being fully built, and if that git dep's
preparation is required for the dependent to properly be installed, oh
well, that's going to break.

But what it WON'T do is create an infinitely deep tree of processes
calling npm and git recursively without end until the user SIGINTs out
of it.

See: https://twitter.com/addaleax/status/1347909082829750273

CC: @addaleax

PR-URL: https://github.com/npm/pacote/pull/64
Credit: @isaacs
Close: #64
Reviewed-by: @ruyadorno